+++ vm/ir/build.h
@@ -26,7 +26,7 @@ vm_ir_arg_t vm_ir_arg_nil(void);
 vm_ir_arg_t vm_ir_arg_bool(bool t);
 vm_ir_arg_t vm_ir_arg_reg(size_t reg);
 vm_ir_arg_t vm_ir_arg_extern(size_t reg);
-vm_ir_arg_t vm_ir_arg_num(ptrdiff_t num);
+vm_ir_arg_t vm_ir_arg_num(double num);
 vm_ir_arg_t vm_ir_arg_str(const char *str);
 vm_ir_arg_t vm_ir_arg_func(vm_ir_block_t *func);
